+++ test/config.js
@@ -1,5 +1,7 @@
 module.exports = {
   nightwatch: {
-    src_folders: ['uitest']
+    src_folders: ['uitest'],
+    output_folder: './reports',
+    html_report_folder: './reports'
   }
 };