+++ history.md
@@ -1,4 +1,13 @@
 # 版本变更
+## 0.4.0 (2018-12-24)
+* [EDIT] 默认配置文件 改为 `yyt.config.js`
+* [ADD] 新增 `yyt path/to/config` 方式
+
+## 0.3.0 (2018-12-12)
+* [ADD] 加入 `nightwatch-html-reporter`
+* [ADD] 新增 `nightwatch.html_report_folder` 配置项
+* [EDIT] 升级 `nightwatch` 到 `1.x` 版本
+
 ## 0.2.0 (2018-12-12)
 * [ADD] 支持自定义 `custom_commands_path`, `custom_assertions_path`